backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-multipart>=0.0.20
openai>=1.59.0
google-genai>=1.0.0
faster-whisper>=1.0.0
qdrant-client>=1.12.0
click>=8.1.0
python-dotenv>=1.0.0
librosa>=0.10.0
scipy>=1.10.0
